% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{fake_data_grein}
\alias{fake_data_grein}
\title{Generates fake dataset starting from breathing support histories taken from
Gren et al.: https://www.nejm.org/doi/full/10.1056/NEJMoa2007016}
\usage{
fake_data_grein()
}
\description{
Generates fake dataset starting from breathing support histories taken from
Gren et al.: https://www.nejm.org/doi/full/10.1056/NEJMoa2007016
}
